refactor index (#230)
* refactor index

* lint fix

* Ignore alto tests

---------
